Isolate process queue
[4.5.x]

ERM37553
