+++ src/IProcessQueue.php
@@ -1,0 +1,67 @@
+<?php
+
+namespace MWStake\MediaWiki\Component\ProcessManager;
+
+use Symfony\Component\Process\Process;
+
+interface IProcessQueue {
+
+	/**
+	 * @param string $pid
+	 * @return ProcessInfo|null if no process found
+	 */
+	public function getProcessInfo( string $pid ): ?ProcessInfo;
+
+	/**
+	 * @param array $steps
+	 * @param int $timeout
+	 * @param array $data
+	 * @return string|null PID of the process or null if process cannot be enqueued
+	 */
+	public function enqueueProcess( array $steps, int $timeout, array $data ): ?string;
+
+	/**
+	 * Proceed with the next step of the process that was previously interrupted
+	 *
+	 * @param string $pid
+	 * @return string|null New PID or null if processes cannot be continued
+	 */
+	public function proceed( $pid ): ?string;
+
+	/**
+	 * @param string $pid
+	 * @param string $lastStep
+	 * @param array $data
+	 * @return bool
+	 */
+	public function recordInterrupt( string $pid, string $lastStep, array $data ): bool;
+
+	/**
+	 * Mark processes as started
+	 *
+	 * @param string $pid
+	 * @return bool
+	 */
+	public function recordStart( string $pid ): bool;
+
+	/**
+	 * @param string $pid
+	 * @param string $step
+	 * @return bool
+	 */
+	public function storeLastCompletedStep( string $pid, string $step ): bool;
+
+	/**
+	 * @param string $pid
+	 * @param int $exitCode
+	 * @param string $exitStatus
+	 * @param array $data
+	 * @return bool
+	 */
+	public function recordFinish( string $pid, int $exitCode, string $exitStatus = '', array $data = [] ): bool;
+
+	/**
+	 * @return ProcessInfo[]
+	 */
+	public function getEnqueuedProcesses(): array;
+}
